Integrating a chatbot with fastAPI/requirements.txt
streamlit
groq
python-dotenv
fastapi
uvicorn
requests